+++ NEWS.md
@@ -1,3 +1,10 @@
+# binfunest 0.2.0
+
+* Added `DBPSKdB` function.
+* Added `marcumq` function (essentially copied from `gsignal` package).
+* Corrected the `DQPDKdB` function.
+* Simplified `B2BConvert`, and improved retention of function call in calls to `stats4::mle`.
+
 # binfunest 0.1.0
 
 * Added a NEWS.md file to track changes to the package.